File change and modification : move exceptions into a dedicated module, move some points board route logic into a function in utils.py, update a generator into list comprehension, add some others minor changes
