+++ flake8-report/index.html
@@ -9,7 +9,7 @@
       <div id="masthead" class="sev-4"></div>
       <div id="page">
           <h1>flake8 violations</h1>
-         <p id="versions">Generated on 2026-03-26 14:43
+         <p id="versions">Generated on 2026-04-02 15:47
             with Installed plugins: flake8-html: 0.4.3, mccabe: 0.7.0, pycodestyle: 2.14.0, pyflakes: 3.4.0
          </p>
          <ul id="index">
@@ -20,7 +20,7 @@
                   <span class="tick">&#x2713;</span>
                </span>
                <h2>All good!</h2>
-               <p>No flake8 errors found in 13 files scanned.</p>
+               <p>No flake8 errors found in 14 files scanned.</p>
             </div>
          </li>
