+++ tests/performance_tests/clubs_tmp.json
@@ -37,38 +37,33 @@
             "points": "4"
         },
         {
-            "name": "Olympic Lifters",
-            "email": "contact@olympiclifters.com",
-            "password": "scrypt:32768:8:1$CO1XjG2GYak7bGad$906886f7afc2f66ac120f12b263e195fd44283a3cf0a39f963722ca25fc62e0909629165ff184ab67e43dfb44b4ca87ec7448ce95bcd896af2d2c32a31ffa604",
+            "name": "Barbell Warriors",
+            "email": "info@barbellwarriors.com",
+            "password": "scrypt:32768:8:1$ZmtTQFYQVRwhl8cK$d4abb1c57ba950bb082a6c963605af00a0e120f7b6d7346d2a9419efad6d7d5a78c7d15a32b57d4fc2eab10aa864977be0a3734c24f69bc6694cd9b224b57545",
+            "points": "15"
+        },
+        {
+            "name": "Simply Lift",
+            "email": "john@simplylift.co",
+            "password": "scrypt:32768:8:1$cYYbKbko75gPN3N3$4883a2360abf24fcc9f8329aedfca168994222748933eff4d89c78e5764f0972fce4c478b5737ece1bd31a3d2a0f1a00586096ffbddb79809b9685eabc5e5f4e",
+            "points": "13"
+        },
+        {
+            "name": "The Weightroom",
+            "email": "team@weightroom.com",
+            "password": "scrypt:32768:8:1$38I8jEaWVYS5jgcA$0b29c9aa0e7f52b2634516d46b312ba429d2748fd4d72973f83275d4ada39907a6bae0389d31a868acdd4b5b43363515186f86e5f7419f9253d092c94d2a9185",
             "points": "15"
         },
         {
             "name": "Iron Titans",
             "email": "contact@irontitans.com",
             "password": "scrypt:32768:8:1$kRW7krsT6zGr4iLk$6a24553114aa57f8cf4297786bf6136fa1d6f5f9832c6157fb406eb4459fcde2a82ed2a64d890ebe5e9a6efe73376239c59a124c77b6bd79b371c84095496b78",
-            "points": "15"
-        },
-        {
-            "name": "Barbell Warriors",
-            "email": "info@barbellwarriors.com",
-            "password": "scrypt:32768:8:1$ZmtTQFYQVRwhl8cK$d4abb1c57ba950bb082a6c963605af00a0e120f7b6d7346d2a9419efad6d7d5a78c7d15a32b57d4fc2eab10aa864977be0a3734c24f69bc6694cd9b224b57545",
             "points": "0",
             "booked_places": {
-                "Heavyweight First": "4",
-                "Powerlifting Open": "4",
-                "Spring Festival": "5",
-                "Regional Strength": "2"
-            }
-        },
-        {
-            "name": "Simply Lift",
-            "email": "john@simplylift.co",
-            "password": "scrypt:32768:8:1$cYYbKbko75gPN3N3$4883a2360abf24fcc9f8329aedfca168994222748933eff4d89c78e5764f0972fce4c478b5737ece1bd31a3d2a0f1a00586096ffbddb79809b9685eabc5e5f4e",
-            "points": "0",
-            "booked_places": {
-                "Heavyweight First": "2",
-                "Spring Festival": "4",
-                "Regional Strength": "6",
+                "Spring Festival": "2",
+                "Regional Strength": "5",
+                "Winter Power": "2",
+                "Heavyweight First": "5",
                 "Powerlifting Open": "1"
             }
         },
@@ -76,25 +71,25 @@
             "name": "She Lifts",
             "email": "kate@shelifts.co.uk",
             "password": "scrypt:32768:8:1$E4fcXT3MQ7oVfc0w$cf31c4a6aec02d4532ffcae03670b12f20265036a3ae5ee941fae6d135ca2db5e406fe9e6c6eded7776e0c256f6d8970f2cbfb74ee50a1c7f9b49b1ad6819b40",
-            "points": "0",
+            "points": "4",
             "booked_places": {
-                "Powerlifting Open": "4",
-                "Spring Festival": "3",
                 "Winter Power": "2",
-                "Regional Strength": "2",
-                "Heavyweight First": "1"
+                "Regional Strength": "1",
+                "Powerlifting Open": "2",
+                "Heavyweight First": "2",
+                "Spring Festival": "1"
             }
         },
         {
-            "name": "The Weightroom",
-            "email": "team@weightroom.com",
-            "password": "scrypt:32768:8:1$38I8jEaWVYS5jgcA$0b29c9aa0e7f52b2634516d46b312ba429d2748fd4d72973f83275d4ada39907a6bae0389d31a868acdd4b5b43363515186f86e5f7419f9253d092c94d2a9185",
-            "points": "4",
+            "name": "Olympic Lifters",
+            "email": "contact@olympiclifters.com",
+            "password": "scrypt:32768:8:1$CO1XjG2GYak7bGad$906886f7afc2f66ac120f12b263e195fd44283a3cf0a39f963722ca25fc62e0909629165ff184ab67e43dfb44b4ca87ec7448ce95bcd896af2d2c32a31ffa604",
+            "points": "9",
             "booked_places": {
-                "Regional Strength": "6",
-                "Winter Power": "2",
-                "Powerlifting Open": "2",
-                "Heavyweight First": "1"
+                "Powerlifting Open": "1",
+                "Regional Strength": "2",
+                "Heavyweight First": "2",
+                "Spring Festival": "1"
             }
         }
     ]
