+++ tests/performance_tests/competitions_tmp.json
@@ -45,25 +45,25 @@
         {
             "name": "Spring Festival",
             "date": "2026-07-27 10:00:00",
-            "number_of_places": "13",
+            "number_of_places": "21",
             "is_past": false
         },
         {
             "name": "Regional Strength",
             "date": "2026-09-12 15:00:00",
-            "number_of_places": "29",
+            "number_of_places": "37",
             "is_past": false
         },
         {
             "name": "Heavyweight First",
             "date": "2026-11-08 13:00:00",
-            "number_of_places": "22",
+            "number_of_places": "21",
             "is_past": false
         },
         {
             "name": "Powerlifting Open",
             "date": "2026-12-05 11:00:00",
-            "number_of_places": "49",
+            "number_of_places": "56",
             "is_past": false
         },
         {
